chore: address review nits for schedule loops
Agent-Logs-Url: https://github.com/rflorian2017/AI-on-the-edge-device/sessions/ed9a5f6b-0c9e-440e-b0b4-5c1d5f96c7b4

+++ code/components/jomjol_flowcontroll/ClassFlowControll.cpp
@@ -602,7 +602,7 @@ bool ClassFlowControll::ReadParameter(FILE* pfile, string& aktparamgraph)
 
             std::vector<std::string> scheduleEntries = ZerlegeZeile(schedule, ";,");
 
-            for (int i = 0; i < scheduleEntries.size(); ++i) {
+            for (size_t i = 0; i < scheduleEntries.size(); ++i) {
                 std::string entry = trim(scheduleEntries[i], " \t");
                 int hour = 0;
                 int minute = 0;
